requirements.txt
langchain
langchain-core
# langchain-embeddings
langchain-community
pymupdf
pypdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
